speed up so the scripts will take half as long to execute for building

+++ scripts/android/build_iconv.sh
@@ -6,7 +6,10 @@ export ICONV_FILE_PATH=$WORKDIR/$ICONV_FILENAME
 export ICONV_SRC_DIR=$WORKDIR/libiconv-1.16
 ICONV_SHA256="e6a1b1b589654277ee790cce3734f07876ac4ccfaecbee8afa0b649cf529cc04"
 
-curl http://ftp.gnu.org/pub/gnu/libiconv/$ICONV_FILENAME -o $ICONV_FILE_PATH
+if [ ! -e "$ICONV_FILE_PATH" ]; then
+  curl http://ftp.gnu.org/pub/gnu/libiconv/$ICONV_FILENAME -o $ICONV_FILE_PATH
+fi
+
 echo $ICONV_SHA256 $ICONV_FILE_PATH | sha256sum -c - || exit 1
 
 for arch in aarch aarch64 i686 x86_64
