+++ scripts/android/build_openssl.sh
@@ -18,7 +18,10 @@ git reset --hard $ZLIB_COMMIT_HASH
 CC=clang CXX=clang++ ./configure --static
 make
 
-curl https://www.openssl.org/source/$OPENSSL_FILENAME -o $OPENSSL_FILE_PATH
+if [ ! -e "$OPENSSL_FILE_PATH" ]; then
+  curl https://www.openssl.org/source/$OPENSSL_FILENAME -o $OPENSSL_FILE_PATH
+fi
+
 echo $OPENSSL_SHA256 $OPENSSL_FILE_PATH | sha256sum -c - || exit 1
 
 for arch in "aarch" "aarch64" "i686" "x86_64"
